requirements.txt
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0

# Optional but HIGHLY RECOMMENDED for semantic verification:
transformers>=4.30.0
torch>=2.0.0

selenium>=4.15.0
webdriver-manager>=4.0.0 

# Optional for DuckDuckGo scraping:
duckduckgo-search>=3.9.0
